fix c++ console color
Refactor code: split LineStyleListener and PositionUpdater in 2
separates classes.

The AnsiConsolePositionUpdater constructs the positions for the escape
codes AND colorized text area. The built AnsiPosition contains the
associated StyleRange.

On the other side, the AnsiConsoleStyleListener initialize the
PositionUpdater when a new document is processed and update the
LyneStyleEvent styles accordingly to the overlapping AnsiPosition.

There was a static AnsiConsoleAttributes variable named current in the
AnsiPosition class that provides random color error when many builds are
run in parallel. This variable does not exist anymore after this
refactoring.

The AnsiConsolePreferenceUtils.tryPreservingStdErrColor() is not
supported by this implementation. I dont't understand the use case.

The offsetToStyleRangeCache does not exist anymore. I haven't seen any
latency in the console with big outputs.